Changed some things
