--- AvalancheWebsiteDocker/web/app/DicordChatExporter.CLI/DiscordChatExporter.Cli.deps.json
@@ -1,163 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v6.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v6.0": {
-      "DiscordChatExporter.Cli/2.33.2": {
-        "dependencies": {
-          "CliFx": "2.2.2",
-          "DiscordChatExporter.Core": "2.33.2",
-          "Gress": "2.0.1",
-          "Spectre.Console": "0.43.0"
-        },
-        "runtime": {
-          "DiscordChatExporter.Cli.dll": {}
-        }
-      },
-      "CliFx/2.2.2": {
-        "runtime": {
-          "lib/netstandard2.1/CliFx.dll": {
-            "assemblyVersion": "2.2.2.0",
-            "fileVersion": "2.2.2.0"
-          }
-        }
-      },
-      "Gress/2.0.1": {
-        "runtime": {
-          "lib/netstandard2.0/Gress.dll": {
-            "assemblyVersion": "2.0.1.0",
-            "fileVersion": "2.0.1.0"
-          }
-        }
-      },
-      "JsonExtensions/1.2.0": {
-        "runtime": {
-          "lib/net5.0/JsonExtensions.dll": {
-            "assemblyVersion": "1.2.0.0",
-            "fileVersion": "1.2.0.0"
-          }
-        }
-      },
-      "MiniRazor.CodeGen/2.2.0": {
-        "dependencies": {
-          "MiniRazor.Runtime": "2.2.0"
-        }
-      },
-      "MiniRazor.Runtime/2.2.0": {
-        "runtime": {
-          "lib/netstandard2.0/MiniRazor.Runtime.dll": {
-            "assemblyVersion": "2.2.0.0",
-            "fileVersion": "2.2.0.0"
-          }
-        }
-      },
-      "Polly/7.2.3": {
-        "runtime": {
-          "lib/netstandard2.0/Polly.dll": {
-            "assemblyVersion": "7.0.0.0",
-            "fileVersion": "7.2.3.0"
-          }
-        }
-      },
-      "Spectre.Console/0.43.0": {
-        "runtime": {
-          "lib/net6.0/Spectre.Console.dll": {
-            "assemblyVersion": "0.0.0.0",
-            "fileVersion": "0.43.0.0"
-          }
-        }
-      },
-      "Superpower/3.0.0": {
-        "runtime": {
-          "lib/net5.0/Superpower.dll": {
-            "assemblyVersion": "1.0.0.0",
-            "fileVersion": "3.0.0.0"
-          }
-        }
-      },
-      "DiscordChatExporter.Core/2.33.2": {
-        "dependencies": {
-          "Gress": "2.0.1",
-          "JsonExtensions": "1.2.0",
-          "MiniRazor.CodeGen": "2.2.0",
-          "Polly": "7.2.3",
-          "Superpower": "3.0.0"
-        },
-        "runtime": {
-          "DiscordChatExporter.Core.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "DiscordChatExporter.Cli/2.33.2": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "CliFx/2.2.2": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-7djhYTJn7VIoMAW9ovOC8vBbJPFWoHo4RocF2iZgjjYO92XZs/Vvn477oi6hIbU4xYcxnNZy/XFJL1xk2+a8eg==",
-      "path": "clifx/2.2.2",
-      "hashPath": "clifx.2.2.2.nupkg.sha512"
-    },
-    "Gress/2.0.1": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-Ky68Wbb3WUkFw5g92tcTRNf5IVMV5AyLB8+uw4pCpXmfS2gkIzWV9BnTGNu+aTaklYqvSZlQL6BnH+9I7HG5tg==",
-      "path": "gress/2.0.1",
-      "hashPath": "gress.2.0.1.nupkg.sha512"
-    },
-    "JsonExtensions/1.2.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-ujtrK6m5BQVQLEteLfl54upN7Z59c1ZqEKmU3OlYriqnV74dIgIlNM2vkcgn2AuZEItQDNORD/2H1biI/asOdg==",
-      "path": "jsonextensions/1.2.0",
-      "hashPath": "jsonextensions.1.2.0.nupkg.sha512"
-    },
-    "MiniRazor.CodeGen/2.2.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-VBV4W4cszTSJBgCk8b9eHMOdsUcumYOGJ09jMK0Wz6M5xO7cmlm8r/QuPIkDMR6+TYM3uT/EVOA8UhbW7BL09g==",
-      "path": "minirazor.codegen/2.2.0",
-      "hashPath": "minirazor.codegen.2.2.0.nupkg.sha512"
-    },
-    "MiniRazor.Runtime/2.2.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-BUrgFWnCA5IsH3rs0MmK1qb/KcvxcoINb4nfPdyt8hp8Z0cVsgv69CKsLPmWhYLgI4BSjFYRsq8OFxy2DYfsXQ==",
-      "path": "minirazor.runtime/2.2.0",
-      "hashPath": "minirazor.runtime.2.2.0.nupkg.sha512"
-    },
-    "Polly/7.2.3": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-DeCY0OFbNdNxsjntr1gTXHJ5pKUwYzp04Er2LLeN3g6pWhffsGuKVfMBLe1lw7x76HrPkLxKEFxBlpRxS2nDEQ==",
-      "path": "polly/7.2.3",
-      "hashPath": "polly.7.2.3.nupkg.sha512"
-    },
-    "Spectre.Console/0.43.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-TyfzVxx3jHr1FeeI9Z6Jn6QTnP7KduV14dn9GTYc0v/t2h4L6WJ0Kc2DOnOJUlGBQ/qXK4P+HITIR/JTHcKYLg==",
-      "path": "spectre.console/0.43.0",
-      "hashPath": "spectre.console.0.43.0.nupkg.sha512"
-    },
-    "Superpower/3.0.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-bjKbAYWePooCAvPxQq+3KnmcLfRggMyRoXmtBMlmCG71bdwBHrO6rmRJ2DRIodg5aztNcxeZXBUVWT+H+MZUhw==",
-      "path": "superpower/3.0.0",
-      "hashPath": "superpower.3.0.0.nupkg.sha512"
-    },
-    "DiscordChatExporter.Core/2.33.2": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
