--- AvalancheWebsiteDocker/web/app/DicordChatExporter.CLI/DiscordChatExporter.Cli.runtimeconfig.json
@@ -1,12 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net6.0",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "6.0.0"
-    },
-    "configProperties": {
-      "System.Reflection.Metadata.MetadataUpdater.IsSupported": false
-    }
-  }
-}
