+++ AvalancheWebsiteDocker/web/nginx.conf
@@ -1,22 +1,13 @@
+events { worker_connections 1024; }
+
 http {
-    include mime.types;
-    sendfile on;
-
     server {
-
-        error_log    /var/log/nginx/SnowtopiaMapError.log debug;
-        rewrite_log on;
-
         listen 8080;
-        listen [::]:8080;
+        root  /web/public;
+        include /etc/nginx/mime.types;
 
-        resolver 127.0.0.1;
-        autoindex off;
-
-        server_name _;
-        server_tokens off;
-
-        root /web/public;
-        gzip_static on;
+        location / {
+            try_files $uri /index.html;
+        }
     }
 }
