+++ AvalancheWebsiteDocker/web/src/MapPage (defualt template).html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en"><head><meta charset="utf-8"/>
+	<meta content="IE-edge" http-equiv="X-UA_Compatible"/>
+	<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
+	<title>Snowtopia Maps</title>
+	<script src="https://code.jquery.com/jquery-3.5.1.js"></script>
+	<script src="https://cdn.datatables.net/1.11.5/js/jquery.dataTables.min.js"></script>
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script> 
+	<link href="https://cdn.datatables.net/v/dt/jq-3.6.0/jszip-2.5.0/dt-1.11.5/af-2.3.7/b-2.2.2/b-colvis-2.2.2/b-html5-2.2.2/b-print-2.2.2/cr-1.5.5/fc-4.0.2/fh-3.2.2/kt-2.6.4/r-2.2.9/rg-1.1.4/rr-1.2.8/sc-2.0.5/sb-1.3.2/sp-2.0.0/sl-1.3.4/sr-1.1.0/datatables.min.css" rel="stylesheet" type="text/css"/>
+ 
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.36/pdfmake.min.js" type="text/javascript"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.1.36/vfs_fonts.js" type="text/javascript"></script>
+	<script src="https://cdn.datatables.net/v/dt/jq-3.6.0/jszip-2.5.0/dt-1.11.5/af-2.3.7/b-2.2.2/b-colvis-2.2.2/b-html5-2.2.2/b-print-2.2.2/cr-1.5.5/fc-4.0.2/fh-3.2.2/kt-2.6.4/r-2.2.9/rg-1.1.4/rr-1.2.8/sc-2.0.5/sb-1.3.2/sp-2.0.0/sl-1.3.4/sr-1.1.0/datatables.min.js" type="text/javascript"></script>
+	<script>
+		$(document).ready(function() {
+		    $('#example').DataTable( {
+		    	"scrollY": "70vh",
+			"scrollCollapse": true,
+		        "paging":   true,
+		        "ordering": true,
+		        "info":     true
+		    } );
+		} );
+	</script>
+</head>
+<body>
+	<h2>Welcome to our collection of community made maps!</h2>
+	<table class="cell-border" id="example" style="width:100%">
+        <thead>
+            <tr>
+                <th>ID</th>
+                <th>Author</th>
+                <th>Content</th>
+                <th>Thumbnail</th>
+                <th>Download</th>
+                <th>Rating</th>
+            </tr>
+        </thead>
+        <tbody>
+		
+        </tbody>
+    </table>
+
+</body></html>
